% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_positions.R
\name{calculate_positions}
\alias{calculate_positions}
\title{Calculate Peptide Positions on a Protein Sequence}
\usage{
calculate_positions(peptide_list, sequence, offset = 0)
}
\arguments{
\item{peptide_list}{A data frame containing a column `Cleaned_Sequence` with peptide sequences.}

\item{sequence}{A string representing the protein sequence.}

\item{offset}{A numeric offset for row numbering.}
}
\value{
A data frame with peptide positions and row numbers.
}
\description{
This function identifies the positions of peptide sequences on a given protein sequence.
It also assigns unique row numbers for visualization.
}
