% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_positions.R
\name{plot_peptide_positions}
\alias{plot_peptide_positions}
\title{Plot Peptide Positions on a Protein Sequence}
\usage{
plot_peptide_positions(
  stroke_positions,
  control_positions,
  protein_length,
  protein_name = "Protein"
)
}
\arguments{
\item{stroke_positions}{Data frame of peptide positions for stroke samples.}

\item{control_positions}{Data frame of peptide positions for control samples.}

\item{protein_length}{The length of the protein sequence.}

\item{protein_name}{The name of the protein (default: "Protein").}
}
\value{
A ggplot2 object showing the peptide positions.
}
\description{
This function visualizes peptide positions on a protein sequence.
}
